fix a bunch of warnings

+++ src/engine/swapChain.hpp
@@ -20,9 +20,9 @@ public:
             swapChain = nullptr;
         }
         for (int i = 0; i < depthImages.size(); i++) {
-            vkDestroyImageView(device.device(), depthImageViews[i], nullptr);
-            vkDestroyImage(device.device(), depthImages[i], nullptr);
-            vkFreeMemory(device.device(), depthImageMemorys[i], nullptr);
+            vkDestroyImageView(device.device(), (unsigned long)depthImageViews[i], nullptr);
+            vkDestroyImage(device.device(), (unsigned long)depthImages[i], nullptr);
+            vkFreeMemory(device.device(), (unsigned long)depthImageMemorys[i], nullptr);
         }
 
         for (auto framebuffer : swapChainFramebuffers) { vkDestroyFramebuffer(device.device(), framebuffer, nullptr); }
@@ -287,11 +287,11 @@ public:
             imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
             imageInfo.flags = 0;
 
-            device.createImageWithInfo(imageInfo, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, depthImages[i], depthImageMemorys[i]);
+            device.createImageWithInfo(imageInfo, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, (unsigned long)depthImages[i], (unsigned long)depthImageMemorys[i]);
 
             VkImageViewCreateInfo viewInfo{};
             viewInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-            viewInfo.image = depthImages[i];
+            viewInfo.image = (unsigned long)depthImages[i];
             viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
             viewInfo.format = depthFormat;
             viewInfo.subresourceRange.aspectMask = VK_IMAGE_ASPECT_DEPTH_BIT;
@@ -300,7 +300,7 @@ public:
             viewInfo.subresourceRange.baseArrayLayer = 0;
             viewInfo.subresourceRange.layerCount = 1;
 
-            if (vkCreateImageView(device.device(), &viewInfo, nullptr, &depthImageViews[i]) != VK_SUCCESS) { throw("failed to create texture image view!"); }
+            if (vkCreateImageView(device.device(), &viewInfo, nullptr, &(unsigned long)depthImageViews[i]) != VK_SUCCESS) { throw("failed to create texture image view!"); }
         }
     }
 
@@ -320,10 +320,10 @@ public:
         for (unsigned int i = 0; i < MAX_FRAMES_IN_FLIGHT; i++) { if (vkCreateSemaphore(device.device(), &semaphoreInfo, nullptr, &imageAvailableSemaphores[i]) != VK_SUCCESS || vkCreateSemaphore(device.device(), &semaphoreInfo, nullptr, &renderFinishedSemaphores[i]) != VK_SUCCESS || vkCreateFence(device.device(), &fenceInfo, nullptr, &inFlightFences[i]) != VK_SUCCESS) { throw("failed to create synchronization objects for a frame!"); }}
     }
 
-    inline VkFramebuffer getFrameBuffer(int index) const { return swapChainFramebuffers[index]; }
+    inline VkFramebuffer getFrameBuffer(unsigned long index) const { return swapChainFramebuffers[index]; }
     inline VkRenderPass getRenderPass() const { return renderPass; }
     inline VkImageView getImageView(int index) const { return swapChainImageViews[index]; }
-    inline unsigned int imageCount() const { return swapChainImages.size(); }
+    inline unsigned long imageCount() const { return swapChainImages.size(); }
     inline VkFormat getSwapChainImageFormat() const { return swapChainImageFormat; }
     inline VkExtent2D getSwapChainExtent() const { return swapChainExtent; }
     inline unsigned int width() const { return swapChainExtent.width; }
